Send a notification

Send a notification via one of the following message/notification apps:

notify_google(msg, config_file = "~/.rgooglespaces", ...)

notify_pushbullet(msg, ...)

notify_slack(msg, config_file = "~/.slackr", ...)

notify_zulip(msg, config_file = "~/.rzulip", ...)

Arguments

msg

the message to send

config_file

character specifying the path to a credentials configuration file

...

additional arguments to internal functions

Value

the response (invisibly)

Details

Google Workspace Spaces

Pushbullet

Slack

Zulip

See below for more details.

Google Workspace Spaces

notify_google() is a wrapper around httr::POST()

Requires a Google Workspace account and 'Space' be setup and configured with a webhook url:

From Gmail, go to 'Spaces' on the left hand navigation bar;

Create a new space or select an existing one;

From the space's dropdown menu, select 'Apps & integrations';

Click on 'Manage webhooks' band create an incoming webhook;

Copy the webhook url;

Create a text file ~/.rgooglespaces, and add the following using your copied webhook url:

webhook_url: https://chat.googleapis.com/v1/spaces/<your_key_and_token_etc>

Pushbullet

notify_pushbullet() is a wrapper around RPushbullet::pbPost()

Use RPushbullet::pbSetup() to configure your machine to use Pushbullet.

Slack

notify_slack() is a wrapper around slackr::slackr_msg()

Zulip

notify_zulip() is a wrapper around httr::POST()

Requires Zulip be setup and configured with a webhook url:

Make note of your Zulip domain (e.g., yourZulipDomain.zulipchat.com)

Create a Zulip bot of type 'Incoming Webhook' following https://zulip.com/help/add-a-bot-or-integration;

Note the bot's API key following https://zulip.com/api/api-keys;

Create a text file ~/.rgooglespaces, and add the following using your bot's API key, your Zulip domain, and optionally a default stream and topic to post notifications to:

zulip_apikey:
zulip_domain: yourZulipDomain.zulipchat.com
zulip_stream: optional stream name
zulip_topic: optional topic name

If a stream is not specified, private messages will be sent to the creator of the bot.